# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Online Smart Question Bank API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /groups:
    get:
      operationId: listGroups
      tags: [academic]
      summary: List groups, optionally filtered by segment
      parameters:
        - name: segmentId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"
...
  /chapters:
    get:
      operationId: listChapters
      tags: [academic]
      summary: List chapters, optionally filtered by subject
      parameters:
        - name: subjectId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of chapters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chapter"
  /topics:
    get:
      operationId: listTopics
      tags: [academic]
      summary: List topics, optionally filtered by chapter
      parameters:
        - name: chapterId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of topics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Topic"
...
  /institutes:
    get:
      operationId: listInstitutes
      tags: [admission]
      summary: List institutes, optionally filtered by admission segment
      parameters:
        - name: admissionSegmentId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of institutes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Institute"
  /admission-exams:
    get:
      operationId: listAdmissionExams
      tags: [admission]
      summary: List admission exams
      parameters:
        - name: admissionSegmentId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of admission exams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdmissionExam"
  /admission-units:
    get:
      operationId: listAdmissionUnits
      tags: [admission]
      summary: List units for a given admission exam
      parameters:
        - name: admissionExamId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of admission units
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdmissionUnit"

  # ── Questions ─────────────────────────────────────────────────────────────
...
  /question-sets:
    get:
      operationId: listQuestionSets
      tags: [question-sets]
      summary: List question sets
      responses:
        "200":
          description: List of question sets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuestionSet"

    post:
      operationId: createQuestionSet
      tags: [question-sets]
      summary: Create a new question set
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionSetInput"
      responses:
        "201":
          description: Created question set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionSet"
  /question-sets/{setId}:
    get:
      operationId: getQuestionSet
      tags: [question-sets]
      summary: Get a question set with its questions
      parameters:
        - name: setId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Question set details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionSetDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteQuestionSet
      tags: [question-sets]
      summary: Delete a question set
      parameters:
        - name: setId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /question-sets/{setId}/items:
    post:
      operationId: addQuestionToSet
      tags: [question-sets]
      summary: Add a question to a set
      parameters:
        - name: setId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionSetItemInput"
      responses:
        "201":
          description: Item added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionSetDetail"
  /question-sets/{setId}/items/{questionId}:
    delete:
      operationId: removeQuestionFromSet
      tags: [question-sets]
      summary: Remove a question from a set
      parameters:
        - name: setId
          in: path
          required: true
          schema:
            type: integer
        - name: questionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # ── Dashboard ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Academic entities ──────────────────────────────────────────────────
...
    Group:
      type: object
      required: [id, segmentId, name]
      properties:
        id:
          type: integer
        segmentId:
          type: integer
        name:
          type: string
        code:
          type: ["string", "null"]
...
    Chapter:
      type: object
      required: [id, subjectId, name, orderNo]
      properties:
        id:
          type: integer
        subjectId:
          type: integer
        name:
          type: string
        orderNo:
          type: integer
    Topic:
      type: object
      required: [id, chapterId, name, orderNo]
      properties:
        id:
          type: integer
        chapterId:
          type: integer
        name:
          type: string
        orderNo:
          type: integer
...
    Institute:
      type: object
      required: [id, admissionSegmentId, name]
      properties:
        id:
          type: integer
        admissionSegmentId:
          type: integer
        name:
          type: string
        shortName:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
    AdmissionExam:
      type: object
      required: [id, admissionSegmentId, name, examYear, examType, createdAt]
      properties:
        id:
          type: integer
        admissionSegmentId:
          type: integer
        name:
          type: string
        examYear:
          type: integer
        conductingBody:
          type: ["string", "null"]
        examType:
          type: string
          enum: [single_institute, cluster, centralized]
        instituteId:
          type: ["integer", "null"]
        negativeMarking:
          type: ["number", "null"]
        createdAt:
          type: string
    AdmissionUnit:
      type: object
      required: [id, admissionExamId, unitName]
      properties:
        id:
          type: integer
        admissionExamId:
          type: integer
        unitName:
          type: string
        description:
          type: ["string", "null"]

    # ── Question entities ──────────────────────────────────────────────────
...
    QuestionSummary:
      type: object
      required: [id, questionType, questionText, status, difficulty, language, createdAt]
      properties:
        id:
          type: integer
        questionType:
          type: string
          enum: [MCQ, CQ, WRITTEN]
        questionText:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        difficulty:
          type: string
          enum: [easy, medium, hard]
        language:
          type: string
          enum: [bn, en]
        subjectName:
          type: ["string", "null"]
        segmentName:
          type: ["string", "null"]
        createdAt:
          type: string
...
    QuestionSet:
      type: object
      required: [id, name, negativeMarking, createdAt, itemCount]
      properties:
        id:
          type: integer
        name:
          type: string
        segmentId:
          type: ["integer", "null"]
        negativeMarking:
          type: number
        createdBy:
          type: ["integer", "null"]
        createdAt:
          type: string
        itemCount:
          type: integer
    QuestionSetDetail:
      type: object
      required: [id, name, negativeMarking, createdAt, items]
      properties:
        id:
          type: integer
        name:
          type: string
        segmentId:
          type: ["integer", "null"]
        negativeMarking:
          type: number
        createdBy:
          type: ["integer", "null"]
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/QuestionSetItem"
    QuestionSetItem:
      type: object
      required: [setId, questionId, orderNo]
      properties:
        setId:
          type: integer
        questionId:
          type: integer
        orderNo:
          type: integer
        marksOverride:
          type: ["number", "null"]
        question:
          $ref: "#/components/schemas/QuestionSummary"
    QuestionSetInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        segmentId:
          type: integer
        negativeMarking:
          type: number
    QuestionSetItemInput:
      type: object
      required: [questionId]
      properties:
        questionId:
          type: integer
        orderNo:
          type: integer
        marksOverride:
          type: number

    # ── Dashboard ──────────────────────────────────────────────────────────
...