# lib/api-zod/src/generated/types/questionSummary.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Online Smart Question Bank API
 * OpenAPI spec version: 0.1.0
 */
import type { QuestionSummaryDifficulty } from './questionSummaryDifficulty';
import type { QuestionSummaryLanguage } from './questionSummaryLanguage';
import type { QuestionSummaryQuestionType } from './questionSummaryQuestionType';
import type { QuestionSummaryStatus } from './questionSummaryStatus';

export interface QuestionSummary {
  id: number;
  questionType: QuestionSummaryQuestionType;
  questionText: string;
  status: QuestionSummaryStatus;
  difficulty: QuestionSummaryDifficulty;
  language: QuestionSummaryLanguage;
  /** @nullable */
  subjectName?: string | null;
  /** @nullable */
  segmentName?: string | null;
  createdAt: string;
}