# lib/api-zod/src/generated/types/topic.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Online Smart Question Bank API
 * OpenAPI spec version: 0.1.0
 */

export interface Topic {
  id: number;
  chapterId: number;
  name: string;
  orderNo: number;
}